Allow Stream to specify custom token width
Currently, the errorItemLength function implicitly assumes that every token
spans exactly one character. That assumption is correct for the Stream
instances for the basic string types (String/Text/ByteString) included in
Megaparsec. It does not necessarily hold for custom token types however,
in particular if using a separate lexer pass.

This commit adds a new tokensLength function that allows the Stream instance
to give a custom width for a non-empty token stream.

In practice, parse errors concerning a token with a width greater than 1
are currently rendered incorrectly (only one caret is displayed below
the offending token).